requirements.txt
django==4.2.11
django-extensions==3.2.3
django-jazzmin==2.6.0
django-debug-toolbar==4.3.0
pylibmc==1.6.3
psycopg2-binary
djangorestframework
drf-yasg